+++ src/App.vue
@@ -4,13 +4,13 @@
     <div class="tab border-1px">
       <div class="tab-item">
         <!-- <router-link> 默认会被渲染成一个 `<a>` 标签 -->
-        <router-link to="/goods">点餐</router-link>
+        <router-link to="/goods"> 点餐 </router-link>
       </div>
       <div class="tab-item">
-        <router-link to="/ratings">评论</router-link>
+        <router-link to="/ratings"> 评论 </router-link>
       </div>
       <div class="tab-item">
-        <router-link to="/seller">商家</router-link>
+        <router-link to="/seller"> 商家 </router-link>
       </div>
     </div>
     <!-- 路由出口:路由匹配到的组件将渲染在这里 -->